Ship-cabin.py, put on a door in a ship's inv
This script should allow players to enter a ship's cabin, and then exit wherever the ship is (even if it moved).  However, in its current state it will generate a new cabin on each use, rather than reusing the map.  This will be changed, but I'm providing this as reference code for other scripts.